mozilla-release/testing/web-platform/tests/service-workers/service-worker/fetch-cors-xhr.https.html

<!DOCTYPE html>
<title>Service Worker: CORS XHR of fetch()</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/get-host-info.sub.js"></script>
<script src="resources/test-helpers.sub.js?pipe=sub"></script>
<body>
<script>
promise_test(function(t) {
    var SCOPE = 'resources/fetch-cors-xhr-iframe.html';
    var SCRIPT = 'resources/fetch-rewrite-worker.js';
    var host_info = get_host_info();

    return login_https(t)
      .then(function() {
          return service_worker_unregister_and_register(t, SCRIPT, SCOPE);
        })
      .then(function(registration) {
          t.add_cleanup(function() {
              return service_worker_unregister(t, SCOPE);
            });

          return wait_for_state(t, registration.installing, 'activated');
        })
      .then(function() { return with_iframe(SCOPE); })
      .then(function(frame) {
          t.add_cleanup(function() {
              frame.remove();
            });

          return new Promise(function(resolve, reject) {
              var channel = new MessageChannel();
<<<<<<< HEAD
              channel.port1.onmessage = resolve;
||||||| merged common ancestors
              channel.port1.onmessage = t.step_func(function(e) {
                  assert_equals(e.data.results, 'finish');
                  frame.remove();
                  service_worker_unregister_and_done(t, SCOPE);
                });
=======
              channel.port1.onmessage = (event) => {
                  if (event.data === 'done') {
                    resolve();
                    return;
                  }
                  test(() => {
                    assert_true(event.data.result);
                  }, event.data.testName);
              };
>>>>>>> upstream-releases
              frame.contentWindow.postMessage({},
                                              host_info['HTTPS_ORIGIN'],
                                              [channel.port2]);
            });
<<<<<<< HEAD
        })
      .then(function(e) {
          assert_equals(e.data.results, 'finish');
        });
||||||| merged common ancestors
        })
      .catch(unreached_rejection(t));
=======
        });
>>>>>>> upstream-releases
  }, 'Verify CORS XHR of fetch() in a Service Worker');
</script>
</body>
